+++ RELEASENOTES.md
@@ -91,6 +91,8 @@
 *   Decoder Extensions (FFmpeg, VP9, AV1, MIDI, etc.):
     *   MIDI decoder: Ignore SysEx event messages
         ([#710](https://github.com/androidx/media/pull/710)).
+    *   MIDI: Fix issue where seeking forward skips the Program Change events
+        ([#704](https://github.com/androidx/media/issues/704).
 *   Leanback extension:
 *   Cast Extension:
     *   Sanitize creation of a `Timeline` to not crash the app when loading